chore: standardize terminology to "Skill atom" for consistency
