+++ app/(auth)/reset-password/component/reset-password-form.tsx
@@ -6,20 +6,16 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Loader } from "lucide-react";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { authApi, ApiError } from "@/lib/api/reset-password";
+import { resetPasswordSchema } from "@/lib/schemas/reset-passord";
 
-const resetPasswordSchema = z.object({
-  email: z
-    .string()
-    .min(1, "Email is required")
-    .email("Please enter a valid email address"),
-});
 
-type ResetPasswordFormData = z.infer<typeof resetPasswordSchema>;
+type ResetPasswordFormData = {
+  email: string;
+};
 
 export function ResetPasswordForm() {
   const router = useRouter();
