+++ app/dashboard/components/top-actions.tsx
@@ -1,7 +1,7 @@
 import { Select } from "@/components/custom/custom-selector";
 import { Input } from "@/components/ui/input";
 import React from "react";
-import { AddUserSheet } from "../user-management/components/add-user-sheet";
+// import { AddUserSheet } from "../user-management/components/add-user-sheet";
 import { AddSkillAtom } from "../skill-atom/components/add-skill-atom";
 
 export const TopActions = () => {
@@ -15,7 +15,7 @@ export const TopActions = () => {
           <Select placeholder="All statuses" options={statusOptions} />
         </div>
       </article>
-      <AddUserSheet />
+      {/* <AddUserSheet /> */}
       <AddSkillAtom />
       
     </section>
