+++ app/dashboard/skill-atom/components/skill-atom-card.tsx
@@ -2,7 +2,7 @@
 
 import { Card, CardContent } from "@/components/ui/card"
 import { FileText } from "lucide-react"
-import { LessonCardMenu } from "./lesson-card-menu"
+import { LessonCardMenu } from "./skill-atom-card-menu"
 
 interface LessonCardProps {
   lesson: {
@@ -15,7 +15,7 @@ interface LessonCardProps {
   onDelete?: (id: number) => void
 }
 
-export const LessonCard = ({ lesson, onView, onEdit, onDelete }: LessonCardProps) => {
+export const SkillAtomCard = ({ lesson, onView, onEdit, onDelete }: LessonCardProps) => {
   return (
     <Card className="hover:shadow-md  cursor-pointer px-3 py-2 group bg-gray-stroke-weak/20">
       <CardContent className="p-1">
