+++ app/dashboard/skill-atom/components/skill-atoms-list.tsx
@@ -1,24 +1,9 @@
-// import Image from "next/image";
-// import React from "react";
 
-// export const SkillAtom = () => {
-//   return (
-//     <section className="w-full mt-4  h-full  flex items-center justify-center flex-col">
-//       <Image
-//         src={"/not-found.png"}
-//         alt="No users found"
-//         height={100}
-//         width={100}
-//       />
-//       No No skill tag added yet
-//     </section>
-//   );
-// };
 "use client"
 
 import React from "react"
 import Image from "next/image"
-import { LessonCard } from "./lesson-card"
+import { SkillAtomCard } from "./skill-atom-card"
 
 interface Lesson {
   id: number
@@ -33,7 +18,7 @@ interface LessonsListProps {
   onDelete?: (id: number) => void
 }
 
-export const LessonsList: React.FC<LessonsListProps> = ({ lessons, onView, onEdit, onDelete }) => {
+export const SkillAtomsList: React.FC<LessonsListProps> = ({ lessons, onView, onEdit, onDelete }) => {
   if (lessons.length === 0) {
     return (
       <section className="flex flex-col items-center justify-center w-full h-full mt-4 min-h-[400px]">
@@ -51,7 +36,7 @@ export const LessonsList: React.FC<LessonsListProps> = ({ lessons, onView, onEdi
   return (
     <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 mt-4">
       {lessons.map((lesson) => (
-        <LessonCard
+        <SkillAtomCard
           key={lesson.id}
           lesson={lesson}
           onView={onView}
