+++ lib/api/skill-atom-dummydata.ts
@@ -1,0 +1,26 @@
+export const dummyLessons = [
+  {
+    id: 1,
+    title: "Introduction to React",
+    subtitle: "Learn about components, props, and state in React.",
+    status: "draft",
+  },
+  {
+    id: 2,
+    title: "Advanced JavaScript",
+    subtitle: "Closures, async/await, and ES6+ features explained.",
+    status: "published",
+  },
+  {
+    id: 3,
+    title: "Next.js Basics",
+    subtitle: "Pages, layouts, routing, and API routes.",
+    status: "published",
+  },
+  {
+    id: 4,
+    title: "TypeScript for Beginners",
+    subtitle: "Strongly type your React apps with TypeScript.",
+    status: "draft",
+  },
+]
